requirements.txt
gradio>=3.50.0
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=0.19.0
